+++ tests/lint/gettickcount-wraparound.pas
@@ -1,0 +1,15 @@
+unit Ticks;
+
+interface
+
+implementation
+
+procedure P;
+var
+  T: Cardinal;
+begin
+  T := GetTickCount;
+  T := GetTickCount64;
+end;
+
+end.