+++ tests/lint/hardcoded-connection-string.pas
@@ -1,0 +1,15 @@
+unit ConnStr;
+
+interface
+
+implementation
+
+procedure P;
+var
+  S: string;
+begin
+  S := 'Data Source=db;User ID=sa;Password=secret';
+  S := 'just a label';
+end;
+
+end.