+++ tests/lint/hardcoded-ip-address.pas
@@ -1,0 +1,15 @@
+unit IPAddr;
+
+interface
+
+implementation
+
+procedure P;
+var
+  S: string;
+begin
+  S := '192.168.1.100';
+  S := 'hello world';
+end;
+
+end.